road to fully automated downloading/generation/correction
