+++ generator/Gemfile
@@ -4,3 +4,5 @@ source "https://rubygems.org"
 git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 
 gem "sxp"
+gem "rake"
+gem "http"
