+++ generator/lib/download.rb
@@ -1,0 +1,36 @@
+require 'http'
+
+class Download
+  PARSE_PATTERNS = {
+    head: '(defcustom dumb-jump-find-rules',
+    tail: '"List of regex patttern templates'
+  }
+
+  def call
+    scp    = HTTP.get('https://raw.githubusercontent.com/jacktasia/dumb-jump/master/dumb-jump.el').to_s
+    result = ""
+    head   = scp.index(PARSE_PATTERNS[:head])
+    tail   = scp.index(PARSE_PATTERNS[:tail])
+
+    # 1. get body
+    extracted = scp[head,tail - head]
+
+    # 2. remove definition start
+    extracted.sub!('(defcustom dumb-jump-find-rules', '')
+
+    # 3. remove whitespaces and some syntax elements
+    extracted.strip!
+    extracted.sub!(/^'\(\(/, '((')
+
+    # 4. remove tabulation
+    extracted.each_line do |line|
+      result << if line[0,2] == "  "
+        line[2, line.size - 2]
+      else
+        line
+      end
+    end
+
+    return result
+  end
+end
